Persist state now detected correctly.
For some utterly bizarre reason, calling param as part of the create_session call can result in create_session
not getting a second argument. At all. it just sees the userid and the savestate hash. Which should be
impossible, but it happens anyway.

This method appears more reliable, and should be more secure too.