Fix for height issues when reusing the same popup window.
If a popup exceeded the height of the window it would get 'height:
<maxheigh>' set on the popup element before being opened. This would
then /not get removed/ when the popup closed - from that point on, the
popup would always be max height, even if the content did not require
it.

This fix sets the height back to auto each time a resize happens, and
the max height is only set explicitly if needed.